Refactor: classes to modules
