+++ src/war_factory.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+from src.building import Building
+from src.constants import Team
+
+
+class WarFactory(Building):
+    COST = 1000
+    POWER_USAGE = 35
+
+    def __init__(self, x: float, y: float, team: Team) -> None:
+        super().__init__(
+            x,
+            y,
+            team,
+            (170, 170, 0) if team == Team.GDI else (170, 0, 0),
+            800,
+        )
